pyproject.toml
[project]
name = "quietreasoning"
version = "0.1.0"
description = "Latent workspace language model with memory routing and TPU-optimized JAX training."
authors = [{ name = "Quiet Reasoning Team" }]
readme = "spec.md"
requires-python = ">=3.10"
dependencies = [
  "jax[tpu]>=0.4.34",
  "flax>=0.8.1",
  "optax>=0.2.3",
  "orbax-checkpoint>=0.5.7",
  "numpy>=1.26.0",
  "sentencepiece>=0.1.99",
  "chex>=0.1.86",
  "einops>=0.7.0",
  "faiss-gpu>=1.8.0",
  "scann>=1.3.0",
  "ml-collections>=0.1.1",
  "tensorflow>=2.16.0",
  "tensorflow-text>=2.16.0"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"quietreasoning" = ["py.typed"]

